Put search direction into own module.
